Restore original test
