+++ radix-engine-tests/tests/logging.rs
@@ -1,121 +1,112 @@
 use radix_engine::{
-    errors::{RuntimeError, SystemModuleError},
-    system::system_modules::limits::TransactionLimitsError,
+    errors::{ApplicationError, RuntimeError},
+    transaction::TransactionReceipt,
     types::*,
 };
+use radix_engine_interface::types::Level;
 use scrypto_unit::*;
 use transaction::builder::ManifestBuilder;
 
-fn prepare_code(message_size: usize, iterations: usize) -> Vec<u8> {
-    wat2wasm(
-        r##"
-(module
-   (import "env" "emit_log" (func $emit_log (param i32 i32 i32 i32)))
-   (data (i32.const 0) "\5C\22\01\00TEXT_CONTENT")
-   (func $test (param $0 i64) (result i64)
-        ;; create a local variable and initialize it to 0
-        (local $i i32)
+fn call<S: AsRef<str>>(function_name: &str, message: S) -> TransactionReceipt {
+    let mut test_runner = TestRunner::builder().build();
+    let package_address = test_runner.compile_and_publish("./tests/blueprints/logger");
 
-        (loop $my_loop
-
-            ;; add one to $i
-            local.get $i
-            i32.const 1
-            i32.add
-            local.set $i
-
-            ;; emit log
-            (call $emit_log
-                (i32.const 0)
-                (i32.const 4)
-                (i32.const 4)
-                (i32.const TEXT_SIZE)
-            )
-
-            ;; if $i is less than ITERATIONS branch to loop
-            local.get $i
-            i32.const ITERATIONS
-            i32.lt_s
-            br_if $my_loop
+    let manifest = ManifestBuilder::new()
+        .call_function(
+            package_address,
+            "Logger",
+            function_name,
+            manifest_args!(message.as_ref().to_owned()),
         )
-        
-        ;; Encode () in SBOR at address 0x0
-        (i32.const 0)
-        (i32.const 92)  ;; prefix
-        (i32.store8)
-        (i32.const 1)
-        (i32.const 33)  ;; tuple value kind
-        (i32.store8)
-        (i32.const 2)
-        (i32.const 0)  ;; tuple length
-        (i32.store8)
-    
-        ;; Return slice (ptr = 0, len = 3)
-        (i64.const 3)
-   )
-   (memory $0 64)
-   (export "memory" (memory $0))
-   (export "Test_f" (func $test))
-)
-    "##
-        .replace("TEXT_CONTENT", " ".repeat(message_size).as_str())
-        .replace("TEXT_SIZE", message_size.to_string().as_str())
-        .replace("ITERATIONS", iterations.to_string().as_str())
-        .as_str(),
-    )
+        .build();
+    let receipt = test_runner.execute_manifest_ignoring_fee(manifest, vec![]);
+
+    receipt
 }
 
-fn test_emit_log(message_size: usize, iterations: usize, expected_err: Option<RuntimeError>) {
+#[test]
+fn test_log_message() {
     // Arrange
-    let code = prepare_code(message_size, iterations);
-    let mut test_runner = TestRunner::builder().without_trace().build();
-    let package_address = test_runner.publish_package(
-        code,
-        single_function_package_definition("Test", "f"),
-        BTreeMap::new(),
-        OwnerRole::None,
-    );
+    let function_name = "log_message";
+    let message = "Hello";
 
     // Act
-    let manifest = ManifestBuilder::new()
-        .lock_fee(test_runner.faucet_component(), 10u32.into())
-        .call_function(package_address, "Test", "f", manifest_args!())
-        .build();
+    let receipt = call(function_name, message);
 
     // Assert
-    let receipt = test_runner.execute_manifest(manifest, vec![]);
-    if let Some(e) = expected_err {
-        receipt.expect_specific_failure(|x| x.eq(&e));
-    } else {
+    {
         receipt.expect_commit_success();
+
+        let logs = receipt.expect_commit(true).application_logs.clone();
+        let expected_logs = vec![(Level::Info, message.to_owned())];
+
+        assert_eq!(expected_logs, logs)
     }
 }
 
 #[test]
-fn test_emit_some_logs() {
-    test_emit_log(DEFAULT_MAX_LOG_SIZE, DEFAULT_MAX_NUMBER_OF_LOGS - 1, None);
+fn test_rust_panic() {
+    // Arrange
+    let function_name = "rust_panic";
+    let message = "Hey";
+
+    // Act
+    let receipt = call(function_name, message);
+
+    // Assert
+    {
+        let logs = receipt.expect_commit(false).application_logs.clone();
+        assert!(logs.is_empty());
+
+        receipt.expect_specific_failure(|e| match e {
+            RuntimeError::ApplicationError(ApplicationError::Panic(e)) => {
+                e.eq("Hey @ logger/src/lib.rs:15:13")
+            }
+            _ => false,
+        })
+    }
 }
 
 #[test]
-fn test_emit_large_logs() {
-    test_emit_log(
-        DEFAULT_MAX_LOG_SIZE + 1,
-        1,
-        Some(RuntimeError::SystemModuleError(
-            SystemModuleError::TransactionLimitsError(TransactionLimitsError::LogSizeTooLarge {
-                actual: DEFAULT_MAX_LOG_SIZE + 1,
-                max: DEFAULT_MAX_LOG_SIZE,
-            }),
-        )),
-    );
+fn test_scrypto_panic() {
+    // Arrange
+    let function_name = "scrypto_panic";
+    let message = "Hi";
+
+    // Act
+    let receipt = call(function_name, message);
+
+    // Assert
+    {
+        let logs = receipt.expect_commit(false).application_logs.clone();
+        assert!(logs.is_empty());
+
+        receipt.expect_specific_failure(|e| match e {
+            RuntimeError::ApplicationError(ApplicationError::Panic(e)) => e.eq(message),
+            _ => false,
+        })
+    }
 }
+
+// FIXME: add length tests when adding event/log/panic message length limit.
+
 #[test]
-fn test_emit_lots_of_logs() {
-    test_emit_log(
-        DEFAULT_MAX_LOG_SIZE,
-        1_000_000,
-        Some(RuntimeError::SystemModuleError(
-            SystemModuleError::TransactionLimitsError(TransactionLimitsError::TooManyLogs),
-        )),
-    );
+fn test_assert_length_5() {
+    // Arrange
+    let function_name = "assert_length_5";
+    let message = "!5";
+
+    // Act
+    let receipt = call(function_name, message);
+
+    // Assert
+    {
+        let logs = receipt.expect_commit(false).application_logs.clone();
+        assert!(logs.is_empty());
+
+        receipt.expect_specific_failure(|e| match e {
+            RuntimeError::ApplicationError(ApplicationError::Panic(e)) => e.eq("assertion failed: `(left == right)`\n  left: `2`,\n right: `5` @ logger/src/lib.rs:23:13"),
+            _ => false,
+        })
+    }
 }
